# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for WeCare Holidays applications",
        "title": "WeCare Holidays API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.wecareholidays.com/support",
            "email": "support@wecareholidays.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/locations": {
            "get": {
                "description": "Get all locations with pagination and filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locations"
                ],
                "summary": "List locations",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by name (partial match, case-insensitive)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by organization type (exact match)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by state (partial match, case-insensitive)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by country (partial match, case-insensitive)",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Filter by tags (exact match, any)",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Filter by aliases (exact match, any)",
                        "name": "aliases",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include soft-deleted locations",
                        "name": "includeDeleted",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with paginated locations",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SwaggerStandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PaginatedLocationsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new location with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locations"
                ],
                "summary": "Create a new location",
                "parameters": [
                    {
                        "description": "Location data",
                        "name": "location",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateLocationDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SwaggerStandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Location"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/organizations": {
            "get": {
                "description": "Get all organizations with pagination and filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizations"
                ],
                "summary": "List organizations",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by name (partial match, case-insensitive)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by slug (partial match, case-insensitive)",
                        "name": "slug",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by organization type (exact match)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (exact match)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by email (partial match, case-insensitive)",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by phone number (partial match, case-insensitive)",
                        "name": "phone",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by website (partial match, case-insensitive)",
                        "name": "website",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by city (partial match, case-insensitive)",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by state (partial match, case-insensitive)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by country (partial match, case-insensitive)",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include soft-deleted organizations",
                        "name": "includeDeleted",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with paginated organizations",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SwaggerStandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.PaginatedOrganizationsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new organization with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizations"
                ],
                "summary": "Create a new organization",
                "parameters": [
                    {
                        "description": "Organization data",
                        "name": "organization",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateOrganizationDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.SwaggerStandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.Organization"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CoordinatesDto": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number",
                    "example": 26.727
                },
                "lng": {
                    "type": "number",
                    "example": 88.395
                }
            }
        },
        "dto.CreateLocationDto": {
            "type": "object",
            "required": [
                "country",
                "district",
                "name",
                "pincode",
                "state",
                "type"
            ],
            "properties": {
                "aliases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "coordinates": {
                    "description": "Optional fields",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.CoordinatesDto"
                        }
                    ]
                },
                "country": {
                    "type": "string",
                    "example": "India"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string",
                    "example": "Darjeeling"
                },
                "geojson": {},
                "mediaUrls": {
                    "$ref": "#/definitions/dto.MediaUrlsDto"
                },
                "name": {
                    "description": "Required fields",
                    "type": "string",
                    "example": "Siliguri"
                },
                "pincode": {
                    "type": "string",
                    "example": "734001"
                },
                "state": {
                    "type": "string",
                    "example": "West Bengal"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.LocationType"
                        }
                    ],
                    "example": "city"
                }
            }
        },
        "dto.CreateOrganizationDto": {
            "type": "object",
            "required": [
                "email",
                "name",
                "type"
            ],
            "properties": {
                "address": {
                    "$ref": "#/definitions/dto.AddressDto"
                },
                "email": {
                    "type": "string",
                    "example": "contact@wecareholidays.com"
                },
                "logo": {
                    "type": "string",
                    "example": "https://storage.example.com/logos/wecare.png"
                },
                "name": {
                    "description": "Required fields",
                    "type": "string",
                    "example": "WeCare Holidays"
                },
                "phone": {
                    "type": "string",
                    "example": "+91 1234567890"
                },
                "slug": {
                    "description": "Optional fields",
                    "type": "string",
                    "example": "wecare-holidays"
                },
                "status": {
                    "description": "If not provided, defaults to \"Pending\"",
                    "type": "string",
                    "example": "Approved"
                },
                "taxIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['GST123456'",
                        " 'PAN1234567']"
                    ]
                },
                "type": {
                    "description": "SUPPLIER, TRAVEL_AGENT, PLATFORM",
                    "type": "string",
                    "example": "SUPPLIER"
                },
                "website": {
                    "type": "string",
                    "example": "https://www.wecareholidays.com"
                }
            }
        },
...
        "dto.LocationType": {
            "type": "string",
            "enum": [
                "city",
                "scenic_spot",
                "transit_hub",
                "district",
                "village",
                "region"
            ],
            "x-enum-varnames": [
                "City",
                "ScenicSpot",
                "TransitHub",
                "District",
                "Village",
                "Region"
            ]
        },
        "dto.MediaUrlsDto": {
            "type": "object",
            "properties": {
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "videos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "dto.PaginatedLocationsResponse": {
            "type": "object",
            "properties": {
                "includeDeleted": {
                    "type": "boolean",
                    "example": false
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Location"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 2
                },
                "totalPages": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.PaginatedOrganizationsResponse": {
            "type": "object",
            "properties": {
                "includeDeleted": {
                    "type": "boolean",
                    "example": false
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Organization"
                    }
                },
                "limit": {
                    "type": "integer",
                    "example": 10
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "total": {
                    "type": "integer",
                    "example": 2
                },
                "totalPages": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "entity.Coordinates": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                }
            }
        },
...
        "entity.Location": {
            "type": "object",
            "properties": {
                "aliases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "coordinates": {
                    "$ref": "#/definitions/entity.Coordinates"
                },
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "createdBy": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "district": {
                    "type": "string"
                },
                "geojson": {},
                "id": {
                    "type": "string"
                },
                "mediaUrls": {
                    "$ref": "#/definitions/entity.MediaURLs"
                },
                "name": {
                    "type": "string"
                },
                "pincode": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "city | scenic_spot | ...",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "entity.MediaURLs": {
            "type": "object",
            "properties": {
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "videos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "entity.Organization": {
            "description": "Organization entity representing clients, suppliers, or travel agents",
            "type": "object",
            "properties": {
                "address": {
                    "description": "Physical address",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Address"
                        }
                    ]
                },
                "createdAt": {
                    "description": "Creation timestamp",
                    "type": "string"
                },
                "deletedAt": {
                    "description": "Soft delete timestamp",
                    "type": "string"
                },
                "email": {
                    "description": "Primary contact email",
                    "type": "string",
                    "example": "contact@wecareholidays.com"
                },
                "id": {
                    "description": "Unique identifier for the organization",
                    "type": "string",
                    "example": "5f8d0c1b7ea3f0d0f3c8e1b9"
                },
                "logo": {
                    "description": "URL to organization logo",
                    "type": "string",
                    "example": "https://storage.example.com/logos/wecare.png"
                },
                "name": {
                    "description": "Organization name",
                    "type": "string",
                    "example": "WeCare Holidays"
                },
                "phone": {
                    "description": "Contact phone number",
                    "type": "string",
                    "example": "+91 1234567890"
                },
                "slug": {
                    "description": "URL-friendly version of name",
                    "type": "string",
                    "example": "wecare-holidays"
                },
                "status": {
                    "description": "Current status (Pending, Approved, Suspended, Archived)",
                    "type": "string",
                    "example": "Approved"
                },
                "taxIds": {
                    "description": "Tax identification numbers",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['GST123456'",
                        " 'PAN1234567']"
                    ]
                },
                "type": {
                    "description": "Organization type (SUPPLIER, TRAVEL_AGENT, PLATFORM)",
                    "type": "string",
                    "example": "SUPPLIER"
                },
                "updatedAt": {
                    "description": "Last update timestamp",
                    "type": "string"
                },
                "website": {
                    "description": "Organization website URL",
                    "type": "string",
                    "example": "https://www.wecareholidays.com"
                }
            }
        },
...
        "models.SwaggerErrorResponse": {
            "description": "Error response model",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Optional error details"
                },
                "errorMessage": {
                    "description": "Human-readable error message",
                    "type": "string",
                    "example": "Invalid request parameters"
                },
                "statusCode": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "description": "Indicates request failure",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwaggerStandardResponse": {
            "description": "Standard API response format",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response payload"
                },
                "errorMessage": {
                    "description": "Error message (only present if success is false)",
                    "type": "string",
                    "example": ""
                },
                "statusCode": {
                    "description": "HTTP status code",
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "description": "Indicates whether the request was successful",
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}